+++ lib/features/auth/auth_checker.dart
@@ -1,6 +1,8 @@
 import 'package:book_adapter/controller/firebase_controller.dart';
 import 'package:book_adapter/features/auth/login_view.dart';
+import 'package:book_adapter/features/widgets/async_value_widget.dart';
 import 'package:book_adapter/features/widgets/loading_view.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
@@ -11,17 +13,11 @@ class AuthChecker extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final userStreamAsyncValue = ref.watch(authStateChangesProvider);
 
-    return userStreamAsyncValue.when(
-      data: (user) {
-        if (user == null) {
-          // Route to login screen
-          return LoginView();
-        }
-        // Route to user screen
-        return child;
-      },
-      loading: (userAsync) => const LoadingView(),
-      error: (e, st, userAsync) => Scaffold(body: Center(child: Text('Error: $e'),),),
+    return AsyncValueWidget<User?>(
+      value: userStreamAsyncValue,
+      data: (data) => data == null
+        ? LoginView()
+        : child,
     );
   }
 }
