+++ lib/features/library/widgets/add_book_button.dart
@@ -1,0 +1,19 @@
+
+
+import 'package:book_adapter/features/library/library_view_controller.dart';
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+class AddBookButton extends ConsumerWidget {
+  const AddBookButton({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final LibraryViewController viewController = ref.watch(libraryViewController.notifier);
+    return IconButton(
+      onPressed: () => viewController.addBooks(context),
+      iconSize: 36,
+      icon: const Icon(Icons.add_rounded)
+    );
+  }
+}
