+++ pubspec.yaml
@@ -38,6 +38,7 @@ dependencies:
   uuid: ^3.0.4
   sticky_headers: ^0.2.0
   flutter_hooks: ^0.18.0
+  sliver_tools: ^0.2.5
 
 dev_dependencies:
   build_runner: ^2.1.2
